Clean up: merge main.js and programs.js
