+++ examples/mapbox-streets/main.js
@@ -7,7 +7,6 @@ import { initSerializer, initGLpaint } from "../../";
 const styleHref = "./streets-v8-noInteractive.json";
 const mapboxToken = "pk.eyJ1IjoiamhlbWJkIiwiYSI6ImNqcHpueHpyZjBlMjAzeG9kNG9oNzI2NTYifQ.K7fqhk2Z2YZ8NIV94M-5nA";
 const tileCoords = { z: 15, x: 7703, y: 13544 };
-// const tileCoords = { z: 13, x: 1310, y: 3166 };
 
 export function main() {
   tileStencil.loadStyle(styleHref, mapboxToken).then(getTile);
@@ -36,11 +35,8 @@ function render(data, style) {
   yawgl.resizeCanvasToDisplaySize(canvas, pixRatio);
   const context = yawgl.initContext(canvas);
 
-  const tileContext = initGLpaint({ 
-    context,
-    framebuffer: { buffer: null, size: canvas },
-    multiTile: false,
-  });
+  const framebuffer = { buffer: null, size: canvas };
+  const tileContext = initGLpaint({ context, framebuffer });
 
   Object.values(data.layers).forEach(tileContext.loadBuffers);
   data.atlas = tileContext.loadAtlas(data.atlas);
