+++ src/preamble.glsl
@@ -10,7 +10,7 @@ vec2 tileToMap(vec2 tilePos) {
 
 vec4 mapToClip(vec2 mapPos, float z) {
   vec2 projected = mapPos * screenScale.xy + vec2(-1.0, 1.0);
-  return vec4(projected, z, 1);
+  return vec4(projected, z, 1.0);
 }
 
 float styleScale(vec2 tilePos) {
